requirements.txt
python-telegram-bot==21.4
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0
imageio==2.35.1
numpy==2.0.1
tqdm==4.66.5
python-dotenv==1.0.1
